requirements.txt
 google-api-python-client 
 google-auth-httplib2 
 google-auth-oauthlib
 py-etherscan-api
 requests
 web3